refactor and fix linting
